Add a basic review queue and photo review interface
This allows you to go through a list of queued images that are still in
the "undecided" state, and for each one choose either to approve or
reject it.  An email will be sent to the original uploading user in
either case.  You can also crop the image before uploading.